trading-api-server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
fastapi-cors==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pandas==2.1.4
loguru==0.7.2 